Show template syntax errors when json parse fails
